backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
langchain>=0.2.0
langchain-openai>=0.1.0
google-generativeai
google-cloud-vision
google-cloud-aiplatform
python-multipart
python-pptx
python-docx
google-cloud-documentai
google-generativeai[adk]
google-adk